refactor(CardParagraph): Rename CardParagraph to Paragraph
