+++ src/components/Sidebar/SidebarProfile/index.tsx
@@ -4,12 +4,12 @@ import { useAuth0 } from '../../../utils/auth/auth0';
 import { useUserState } from '../../../utils/hooks/useUserState';
 
 import CardHeadline from '../../Card/CardHeadline/';
-import CardParagraph from '../../Card/CardParagraph/';
+import Paragraph from '../../Paragraph';
 import WateredTreesIndicator from '../../WateredTreesIndicator';
 import CardAccordion from '../../Card/CardAccordion/';
 import CardCredentials from '../../Card/CardCredentials/';
 import TreesAdopted from '../../Card/CardAccordion/TreesAdopted';
-import { NonVerfiedMailCardParagraph } from '../../Card/non-verified-mail';
+import { NonVerfiedMailMessage } from '../../Card/non-verified-mail';
 import Login from '../../Login';
 import ButtonRound from '../../ButtonRound';
 import LoadingIcon from '../../LoadingIcon/';
@@ -65,11 +65,11 @@ const SidebarProfile: FC = () => {
       <>
         <SidebarTitle>Profil</SidebarTitle>
         <FlexCol>
-          <CardParagraph>
+          <Paragraph>
             Du bist momentan nicht eingeloggt. Wenn du das Gießen von Bäumen in
             deiner Umgebung hier eintragen möchtest, dann registriere dich oder
             logge dich ein.
-          </CardParagraph>
+          </Paragraph>
           <Login width='-webkit-fill-available' />
           <ParticipateButton />
         </FlexCol>
@@ -81,7 +81,7 @@ const SidebarProfile: FC = () => {
     return (
       <>
         <SidebarTitle>Profil</SidebarTitle>
-        <NonVerfiedMailCardParagraph />
+        <NonVerfiedMailMessage />
       </>
     );
   }
@@ -97,12 +97,12 @@ const SidebarProfile: FC = () => {
       <CardCredentials />
       <Login width='-webkit-fill-available' />
       <>
-        <CardParagraph>
+        <Paragraph>
           Möchtest du deinen Account löschen? Damit werden alle von dir
           generierten Wässerungsdaten einem anonymen Benutzer zugeordnet. Dein
           Benutzer bei unserem Authentifizierungsdienst Auth0.com wird sofort
           und unwiderruflich gelöscht.
-        </CardParagraph>
+        </Paragraph>
         <LastButtonRound
           width='-webkit-fill-available'
           onClick={evt => {
